delete what previously made

+++ 3-quick_sort.c
@@ -64,7 +64,7 @@ void quick_sort_recursive(int *array, int low, int high, size_t size)
 
 void quick_sort(int *array, size_t size)
 {
-	int low = 0, high = (int)size - 1;
+	int low = 0, high = size - 1;
 
 	if (array == NULL || size < 2)
 		return;
